Reduce GC pressure thresholds in tests
The .NET Framework fix to reduce allocations from use of AsyncLocal<T> allows us to reduce levels back to (or very near to) those set before commit b30ea88fe.

If this fix ships as a servicing release to .NET 4.7.2, an opt-in is required to get the benefit, which is included in this commit.